requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
python-i18n==0.3.9
requests==2.27.1
Unidecode==1.3.2
urllib3==1.26.8
